+++ lib/data_mapper/mapper/builder/relationship/collection_behavior.rb
@@ -1,0 +1,15 @@
+module DataMapper
+  class Mapper
+    class Builder
+      class Relationship
+
+        module CollectionBehavior
+
+          def target_model_attribute_options
+            super.merge(:collection => true)
+          end
+        end # module CollectionBehavior
+      end # class Relationship
+    end # class Builder
+  end # class Mapper
+end # module DataMapper
