+++ lib/data_mapper/mapper/builder/relationship/one_to_many.rb
@@ -1,0 +1,12 @@
+module DataMapper
+  class Mapper
+    class Builder
+      class Relationship
+
+        class OneToMany < OneToOne
+          include CollectionBehavior
+        end # class OneToMany
+      end # class Relationship
+    end # class Builder
+  end # class Mapper
+end # module DataMapper
