--- lib/data_mapper/relationship/builder/collection_behavior.rb
@@ -1,13 +1,0 @@
-module DataMapper
-  class Relationship
-    class Builder
-
-      module CollectionBehavior
-
-        def target_model_attribute_options
-          super.merge(:collection => true)
-        end
-      end # module CollectionBehavior
-    end # class Builder
-  end # class Relationship
-end # module DataMapper
