--- lib/data_mapper/relationship/builder/one_to_many.rb
@@ -1,10 +1,0 @@
-module DataMapper
-  class Relationship
-    class Builder
-
-      class OneToMany < OneToOne
-        include CollectionBehavior
-      end # class OneToMany
-    end # class Builder
-  end # class Relationship
-end # module DataMapper
