+++ lib/dm-mapper.rb
@@ -84,6 +84,13 @@ require 'data_mapper/mapper'
 require 'data_mapper/mapper/relation'
 require 'data_mapper/mapper/relation/base'
 
+require 'data_mapper/mapper/builder/relationship'
+require 'data_mapper/mapper/builder/relationship/collection_behavior'
+require 'data_mapper/mapper/builder/relationship/one_to_one'
+require 'data_mapper/mapper/builder/relationship/one_to_many'
+require 'data_mapper/mapper/builder/relationship/many_to_one'
+require 'data_mapper/mapper/builder/relationship/many_to_many'
+
 require 'data_mapper/relationship'
 require 'data_mapper/relationship/options'
 require 'data_mapper/relationship/options/one_to_one'
@@ -99,11 +106,5 @@ require 'data_mapper/relationship/one_to_one'
 require 'data_mapper/relationship/one_to_many'
 require 'data_mapper/relationship/many_to_one'
 require 'data_mapper/relationship/many_to_many'
-require 'data_mapper/relationship/builder'
-require 'data_mapper/relationship/builder/collection_behavior'
-require 'data_mapper/relationship/builder/one_to_one'
-require 'data_mapper/relationship/builder/one_to_many'
-require 'data_mapper/relationship/builder/many_to_one'
-require 'data_mapper/relationship/builder/many_to_many'
 
 require 'data_mapper/query'
